Refactor useBackPress to show discard modal only if photos taken during this instance of the camera
